# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Rescan API
  version: 0.1.0
  description: Batch uploads and asynchronous Tika/ViT resume processing. All operations require an owner-scoped access token.
servers:
  - url: http://localhost:8080
security:
  - bearer: []
paths:
  /v1/jobs:
    post:
      operationId: createJob
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: {type: string, minLength: 1, maxLength: 128}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  minItems: 1
                  maxItems: 4000
                  items:
                    type: object
                    required: [filename, sizeBytes]
                    properties:
                      filename: {type: string, minLength: 1, maxLength: 255}
                      sizeBytes: {type: integer, format: int64, minimum: 1, maximum: 26214400}
      responses:
        '201':
          description: Job created or existing idempotent job returned; upload page omitted if already submitted.
          content:
            application/json:
              schema:
                allOf:
                  - {$ref: '#/components/schemas/UploadPage'}
                  - type: object
                    properties:
                      jobId: {type: string, format: uuid}
                      total: {type: integer, maximum: 4000}
        '400': {$ref: '#/components/responses/Invalid'}
        '409': {$ref: '#/components/responses/Conflict'}
    get:
      operationId: listJobs
      parameters:
        - {$ref: '#/components/parameters/After'}
        - {$ref: '#/components/parameters/Limit'}
        - in: query
          name: status
          schema: {$ref: '#/components/schemas/JobStatus'}
      responses:
        '200':
          description: Owner's jobs ordered by UUID; items and nextCursor.
          content: {application/json: {schema: {$ref: '#/components/schemas/Page'}}}
  /v1/jobs/{id}:
    parameters:
      - {$ref: '#/components/parameters/JobId'}
    get:
      operationId: getJob
      responses:
        '200':
          description: Job status, total, created_at, updated_at, verified uploaded count, and per-state counts.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: {type: string, format: uuid}
                  status: {$ref: '#/components/schemas/JobStatus'}
                  total: {type: integer}
                  uploaded: {type: integer, description: Number verified at submission, not live S3 upload progress.}
                  counts: {type: object, additionalProperties: {type: integer}}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      operationId: deleteJob
      responses:
        '202': {description: Deletion requested; jobId and DELETING status returned.}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/jobs/{id}/upload-urls:
    parameters:
      - {$ref: '#/components/parameters/JobId'}
    post:
      operationId: uploadUrls
      parameters:
        - {$ref: '#/components/parameters/After'}
        - in: query
          name: limit
          schema: {type: integer, minimum: 1, maximum: 100, default: 100}
      responses:
        '200':
          description: Upload capabilities valid for 15 minutes, with original manifest fileIndex.
          content: {application/json: {schema: {$ref: '#/components/schemas/UploadPage'}}}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /v1/jobs/{id}/submit:
    parameters:
      - {$ref: '#/components/parameters/JobId'}
    post:
      operationId: submitJob
      responses:
        '202': {description: Verification accepted asynchronously, or current submitted job returned. Response contains jobId and status. VERIFYING transitions to QUEUED when every upload is valid, or UPLOADING with document error_code values when correction is needed.}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /v1/jobs/{id}/documents:
    parameters:
      - {$ref: '#/components/parameters/JobId'}
    get:
      operationId: listDocuments
      parameters:
        - {$ref: '#/components/parameters/After'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: Items contain id, filename, size_bytes, status, attempts and error_code; nextCursor is null on the final page.
          content: {application/json: {schema: {$ref: '#/components/schemas/Page'}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/jobs/{id}/documents/{documentId}/result:
    parameters:
      - {$ref: '#/components/parameters/JobId'}
      - in: path
        name: documentId
        required: true
        schema: {type: string, format: uuid}
    get:
      operationId: getResult
      responses:
        '200':
          description: Download the JSON at url within five minutes; content follows result.schema.json.
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: {type: string, format: uri}
                  expiresInSeconds: {type: integer, example: 300}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
components:
  securitySchemes:
    bearer: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    JobId:
      in: path
      name: id
      required: true
      schema: {type: string, format: uuid}
    After:
      in: query
      name: after
      schema: {type: string, format: uuid}
    Limit:
      in: query
      name: limit
      schema: {type: integer, minimum: 1, maximum: 100, default: 50}
  schemas:
    JobStatus:
      type: string
      enum: [UPLOADING, VERIFYING, QUEUED, PROCESSING, SUCCEEDED, PARTIAL_SUCCESS, FAILED, DELETING]
    Page:
      type: object
      properties:
        items: {type: array, items: {type: object}}
        nextCursor: {type: string, format: uuid, nullable: true}
    UploadPage:
      type: object
      properties:
        nextCursor: {type: string, format: uuid, nullable: true}
        uploads:
          type: array
          maxItems: 100
          items:
            type: object
            properties:
              documentId: {type: string, format: uuid}
              fileIndex: {type: integer, minimum: 0, maximum: 3999}
              filename: {type: string}
              method: {type: string, enum: [PUT]}
              url: {type: string, format: uri}
              headers: {type: object, additionalProperties: {type: array, items: {type: string}}}
              expiresInSeconds: {type: integer, example: 900}
  responses:
    Invalid: {description: Invalid manifest, pagination, or configured limit exceeded.}
    Conflict: {description: Conflicting idempotency key, incomplete uploads, invalid job state, or result not ready.}
    NotFound: {description: Missing resource or resource owned by another user.}